refactor: moved access request to separate client
